requirements.txt
aiohttp==3.8.1
Flask==2.1.0
langchain-openai==0.1.0
asyncio
python-dotenv==0.21.0